refactor: rename SignInPage.tsx to AuthPage.tsx (#29)
